fix: make perf-self-profile compile on macOS
Cfg-gates linux-specific code to allow compilation on non-linux platofrms.
Also fixes missing required-features for blocking_pool_tracking example and silences dead code warnings that fired without the cpu-profiling feature enabled.
