+++ perf-self-profile/src/sys.rs
@@ -1,0 +1,9 @@
+#[cfg(target_os = "linux")]
+mod linux;
+#[cfg(target_os = "linux")]
+pub use linux::*;
+
+#[cfg(not(target_os = "linux"))]
+mod unsupported;
+#[cfg(not(target_os = "linux"))]
+pub use unsupported::*;
