+++ perf-self-profile/tests/sched_triggers.rs
@@ -1,3 +1,5 @@
+#![cfg(target_os = "linux")]
+
 use dial9_perf_self_profile::{EventSource, PerfSampler, SamplerConfig, resolve_symbol};
 use std::sync::{Arc, Mutex};
 use std::thread;
